requirements.txt
fastapi
uvicorn
python-dotenv
pyjwt[crypto]
pydantic
pydantic-settings
httpx
openai